Attempt to use gjson syntax to filer for merged prs
I was hoping this field was only present when the PR is merged but of
course it's always there, so i'm trying to use
https://github.com/tidwall/gjson#path-syntax
to filter for only when it's true (even tho i think we're about to
switch to json path).

I feel like filtering is a very different action from parsing, so i hope
we end up using 2 different ways of expressing that instead of mixing
them. I guess there's always gonna be some overlap.